+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 - Initial version, created by Stagehand
+- Initial Dart Server Setup
+- Static Files and README Updates
+- Static File and HTML Handling
+- Setup Dart Workflow (#4)
+- Code Organization